Add RP2040 boot.py that allows CoCo to boot
If the Pico board is powered from a USB power brick rather than connected to my
dev laptop, the RP2040 will come up with defaults that interfere with the
operation of the CoCo. This boot.py file will configure the pin directions on
the RP2040 such that the CoCo can function normally with the CoPiCo board
installed even if the full firmware is not installed on the RP2040.